backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
paho-mqtt==2.1.0
pydantic-settings==2.6.1
python-multipart==0.0.17
alembic==1.14.0
